master와 dev 의 싱크를 맞추기 위해 pull request 합니다

+++ BE/src/main/java/com/codesquad/team10/todo/api/CardController.java
@@ -1,21 +1,24 @@
 package com.codesquad.team10.todo.api;
 
+import com.auth0.jwt.exceptions.JWTDecodeException;
+import com.auth0.jwt.exceptions.SignatureVerificationException;
 import com.codesquad.team10.todo.entity.*;
-import com.codesquad.team10.todo.exception.custom.InvalidRequestException;
-import com.codesquad.team10.todo.exception.custom.ResourceNotFoundException;
-import com.codesquad.team10.todo.exception.custom.UnmatchedRequestDataException;
+import com.codesquad.team10.todo.exception.custom.*;
 import com.codesquad.team10.todo.repository.CardRepository;
 import com.codesquad.team10.todo.repository.LogRepository;
 import com.codesquad.team10.todo.repository.SectionRepository;
 import com.codesquad.team10.todo.response.ResponseData;
+import com.codesquad.team10.todo.util.JWTUtils;
 import com.codesquad.team10.todo.util.ModelMapper;
-import com.codesquad.team10.todo.vo.CardDTO;
+import com.codesquad.team10.todo.dto.CardDTO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -25,10 +28,6 @@ import java.util.Map;
 public class CardController {
 
     private static final Logger logger = LoggerFactory.getLogger(CardController.class);
-
-    private static final Integer TEST_BOARD_ID = 1;
-    private static final String TEST_USER_NAME = "nigayo";
-    private static final Integer TEST_USER_ID = 1;
 
     private SectionRepository sectionRepository;
     private CardRepository cardRepository;
@@ -41,9 +40,17 @@ public class CardController {
     }
 
     @PostMapping("")
-    public ResponseEntity<ResponseData> create(@PathVariable int sectionId, @RequestBody Map<String, String> body) {
+    public ResponseEntity<ResponseData> create(@PathVariable int sectionId, @RequestBody Map<String, String> body, HttpServletRequest request) {
+        User userData = null;
+        try {
+            userData = JWTUtils.getUserFromJWT(request.getHeader(HttpHeaders.AUTHORIZATION));
+        } catch (SignatureVerificationException | NullPointerException e) {
+            throw new ForbiddenException();
+        } catch (JWTDecodeException e) {
+            throw new InvalidTokenException();
+        }
         Section section = sectionRepository.findById(sectionId).orElseThrow(ResourceNotFoundException::new);
-        Card newCard = new Card(body.get("title"), body.get("content"), TEST_USER_NAME, TEST_USER_ID);
+        Card newCard = new Card(body.get("title"), body.get("content"), userData.getName(), userData.getId());
         section.addCard(newCard);
 
         logger.debug("new card: {}", newCard);
@@ -55,7 +62,7 @@ public class CardController {
         CardDTO cardDTO = null;
         cardDTO = (CardDTO) ModelMapper.of(newCard);
         // 로그 추가
-        Log log = new Log(TEST_USER_NAME, Action.ADDED, Target.CARD, newCard.getTitle(), newCard.getContent(), null, section.getTitle(), TEST_BOARD_ID);
+        Log log = new Log(userData.getName(), Action.ADDED, Target.CARD, newCard.getTitle(), newCard.getContent(), null, section.getTitle(), userData.getBoard());
         logRepository.save(log);
         // 반환 데이터
         Map<String, Object> responseData = constructResonseData(cardDTO, log, getCountOfCardWithoutDeleted(section.getCards()));
@@ -63,9 +70,18 @@ public class CardController {
     }
 
     @PatchMapping("/{cardId}")
-    public ResponseEntity<ResponseData> update(@PathVariable int sectionId, @PathVariable int cardId, @RequestBody Map<String, String> body) {
-        if (body.get("content") == null)
+    public ResponseEntity<ResponseData> update(@PathVariable int sectionId, @PathVariable int cardId, @RequestBody Map<String, String> body, HttpServletRequest request) {
+        if (body.get("content") == null) {
             throw new InvalidRequestException();
+        }
+        User userData = null;
+        try {
+            userData = JWTUtils.getUserFromJWT(request.getHeader(HttpHeaders.AUTHORIZATION));
+        } catch (SignatureVerificationException | NullPointerException e) {
+            throw new ForbiddenException();
+        } catch (JWTDecodeException e) {
+            throw new InvalidTokenException();
+        }
         Card updateCard = cardRepository.findById(cardId).orElseThrow(ResourceNotFoundException::new);
         Section section = sectionRepository.findById(sectionId).orElseThrow(ResourceNotFoundException::new);
         if (section.getCards().size() == 0)
@@ -75,7 +91,7 @@ public class CardController {
         CardDTO resultCard = (CardDTO) ModelMapper.of(section.updateCard(updateCard, body.get("title"), body.get("content")));
         sectionRepository.save(section);
         // 로그 추가
-        Log log = new Log(TEST_USER_NAME, Action.UPDATED, Target.CARD, resultCard.getTitle(), resultCard.getContent(), null, null, TEST_BOARD_ID);
+        Log log = new Log(userData.getName(), Action.UPDATED, Target.CARD, resultCard.getTitle(), resultCard.getContent(), null, null, userData.getBoard());
         logRepository.save(log);
         logger.debug("log: {}", log);
         //반환 데이터
@@ -84,15 +100,26 @@ public class CardController {
     }
 
     @DeleteMapping("/{cardId}")
-    public ResponseEntity<ResponseData> delete(@PathVariable int sectionId, @PathVariable int cardId) {
+    public ResponseEntity<ResponseData> delete(@PathVariable int sectionId, @PathVariable int cardId, HttpServletRequest request) {
+        User userData = null;
+        try {
+            userData = JWTUtils.getUserFromJWT(request.getHeader(HttpHeaders.AUTHORIZATION));
+        } catch (SignatureVerificationException | NullPointerException e) {
+            throw new ForbiddenException();
+        } catch (JWTDecodeException e) {
+            throw new InvalidTokenException();
+        }
         Card targetCard = cardRepository.findById(cardId).orElseThrow(ResourceNotFoundException::new);
+        if (targetCard.isDeleted())
+            throw new ResourceNotFoundException();
+
         Section section = sectionRepository.findById(sectionId).orElseThrow(ResourceNotFoundException::new);
         if (section.getCards().size() == 0)
             throw new UnmatchedRequestDataException();
 
         section.deleteCard(targetCard);
         sectionRepository.save(section);
-        Log log = new Log(TEST_USER_NAME, Action.REMOVED, Target.CARD, targetCard.getTitle(), targetCard.getContent(), section.getTitle(), null, TEST_BOARD_ID);
+        Log log = new Log(userData.getName(), Action.REMOVED, Target.CARD, targetCard.getTitle(), targetCard.getContent(), section.getTitle(), null, userData.getBoard());
         logRepository.save(log);
         logger.debug("log: {}", log);
         Map<String, Object> responseData = constructResonseData(log, getCountOfCardWithoutDeleted(section.getCards()));
